Rad na user management sustavu. Potrebno jos popraviti "modify user" opciju.

+++ utils/DBUtils.py
@@ -1,4 +1,5 @@
 from UserManagementApp import db
+from datasource.entity.UserType import UserTypeEnum
 
 class DBUtils:
 
@@ -28,11 +29,23 @@ class DBUtils:
     def findById(clazz, id):
         return clazz.query.filter_by(id=id).one_or_none()
 
+    @staticmethod
+    def findByName(clazz, name):
+        return clazz.query.filter_by(username=name).one_or_none()
+
 
     @staticmethod
     def deleteById(clazz, id):
         model = DBUtils.findById(clazz, id)
-        if model is not None:
+        if model is not None and model.user_type != UserTypeEnum.ADMIN:
+            return DBUtils.delete(model)
+        else:
+            return False
+
+    @staticmethod
+    def deleteByName(clazz, name):
+        model = DBUtils.findByName(clazz, name)
+        if model is not None and model.user_type != UserTypeEnum.ADMIN:
             return DBUtils.delete(model)
         else:
             return False
